requirements.txt
aniso8601==8.0.0
attrs==19.3.0
click==7.1.1
Flask==1.1.1
flask-restplus==0.13.0
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
numpy==1.18.2
Pillow==7.0.0
pyrsistent==0.15.7
pytz==2019.3
six==1.14.0
Werkzeug==0.16.0
zipp==3.1.0
